+++ cbse-blog-server/src/main/java/com/loloao/mapper/ArticleMapper.java
@@ -29,10 +29,10 @@ public interface ArticleMapper extends BaseMapper<Article> {
     int getCountArticleByTagId(Integer id);*/
 
     @Select("SELECT COUNT(*) AS count FROM me_article a WHERE a.category_id = #{id}")
-    int getCountArticleByCategoryId(Integer id);
+    int getCountArticleByCategoryId(Long id);
 
     @Select("SELECT COUNT(*) AS count FROM me_article_tag at WHERE at.tag_id = #{id}")
-    int getCountArticleByTagId(Integer id);
+    int getCountArticleByTagId(Long id);
 
     @Select("SELECT at.article_id FROM me_article_tag at WHERE at.tag_id = #{tagId}")
     List<Integer> getArticleIdsByTagId(Long tagId);
